Remove passing of asyncio event loop as argument.
This removes most cases of passing the asyncio event loop as an argument
(some cases are kept in the Linux backend for interop with txdbus/twisted).

Passing the loop as an argument has been deprecated in Python 3.8 and
will be removed in 3.10.

Explanation: https://stackoverflow.com/a/60315290/1976323

Also, the current implementation was causing problems because
`asyncio.get_event_loop()` was called in `__init__` which could
occur before the event loop is running which gets a different event
loop from the one actually runs.